Refactored the React Files by Adding the Components and Container Folders
