+++ reactProjects/robofriends/src/components/CardList/CardList.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Card from './Card';
+import Card from '../Cards/Card';
 
 const CardList = ({robots})=>{
